HDPI-4268: Move shouldShowStep function into index.ts
